requirements.txt
Django
psycopg2
coreapi
djangorestframework
markdown
django-filter
django-crispy-forms
django-guardian
django-rest-swagger
sphinx
django-autofixture
django-extensions
jupyter
django-debug-toolbar